#130 RoleService cleanup

+++ src/FluentCMS.Services/RoleService.cs
@@ -13,19 +13,17 @@ public interface IRoleService
 }
 
 
-public class RoleService : IRoleService
+public class RoleService : BaseService<Role>, IRoleService
 {
     protected readonly RoleManager<Role> RoleManager;
 
-    public RoleService(RoleManager<Role> roleManager)
+    public RoleService(RoleManager<Role> roleManager, IApplicationContext appContext) : base(appContext)
     {
         RoleManager = roleManager;
     }
 
     public Task<Role> GetById(Guid id, CancellationToken cancellationToken = default)
     {
-        cancellationToken.ThrowIfCancellationRequested();
-
         var roles = RoleManager.Roles.AsEnumerable().Single(r => r.Id.Equals(id));
 
         return Task.FromResult(roles);
@@ -33,8 +31,6 @@ public class RoleService : IRoleService
 
     public Task<IEnumerable<Role>> GetAll(Guid siteId, CancellationToken cancellationToken = default)
     {
-        cancellationToken.ThrowIfCancellationRequested();
-
         var roles = RoleManager.Roles.Where(role => role.SiteId.Equals(siteId)).AsEnumerable();
 
         return Task.FromResult(roles);
@@ -42,7 +38,13 @@ public class RoleService : IRoleService
 
     public async Task<Role> Create(Role role, CancellationToken cancellationToken)
     {
-        cancellationToken.ThrowIfCancellationRequested();
+        if (role.SiteId != Current.Site.Id)
+            throw new Exception("Role must be created for the current site.");
+
+        if (!Current.IsInRole(Current.Site.AdminRoleIds))
+            throw new Exception("Only admin can create a role.");
+
+        PrepareForCreate(role);
 
         var idResult = await RoleManager.CreateAsync(role);
 
@@ -53,7 +55,13 @@ public class RoleService : IRoleService
 
     public async Task<Role> Update(Role role, CancellationToken cancellationToken)
     {
-        cancellationToken.ThrowIfCancellationRequested();
+        if (role.SiteId != Current.Site.Id)
+            throw new Exception("Role must be updated for the current site.");
+
+        if (!Current.IsInRole(Current.Site.AdminRoleIds))
+            throw new Exception("Only admin can update a role.");
+
+        PrepareForUpdate(role);
 
         var idResult = await RoleManager.UpdateAsync(role);
 
@@ -64,7 +72,11 @@ public class RoleService : IRoleService
 
     public async Task Delete(Role role, CancellationToken cancellationToken = default)
     {
-        cancellationToken.ThrowIfCancellationRequested();
+        if (role.SiteId != Current.Site.Id)
+            throw new Exception("Role must be deleted for the current site.");
+
+        if (!Current.IsInRole(Current.Site.AdminRoleIds))
+            throw new Exception("Only admin can update a role.");
 
         var idResult = await RoleManager.DeleteAsync(role);
